enhance: Extract EntitySchema type helpers into own file
