+++ packages/endpoint/src/schemas/EntityTypes.ts
@@ -1,57 +1,6 @@
-// we just removed instances of 'abstract new'
-import type { Schema, GetIndex, GetEntity, CheckLoop } from '../interface.js';
+import type { Schema, GetEntity, GetIndex } from '../interface.js';
 import { AbstractInstanceType } from '../normal.js';
-export type Constructor = new (...args: any[]) => {};
-export type IDClass = new (...args: any[]) => {
-  id: string | number | undefined;
-};
-export type PKClass = new (...args: any[]) => {
-  pk(
-    parent?: any,
-    key?: string,
-    args?: readonly any[],
-  ): string | number | undefined;
-};
-type ValidSchemas<TInstance> = {
-  [k in keyof TInstance]?: Schema;
-};
-export type EntityOptions<TInstance extends {}> = {
-  readonly schema?: ValidSchemas<TInstance>;
-  readonly pk?:
-    | ((
-        value: TInstance,
-        parent?: any,
-        key?: string,
-      ) => string | number | undefined)
-    | keyof TInstance;
-  readonly key?: string;
-} & {
-  readonly [K in Extract<
-    keyof IEntityClass,
-    | 'process'
-    | 'merge'
-    | 'expiresAt'
-    | 'createIfValid'
-    | 'mergeWithStore'
-    | 'validate'
-    | 'shouldReorder'
-    | 'shouldUpdate'
-  >]?: IEntityClass<new (...args: any[]) => TInstance>[K];
-};
-export interface RequiredPKOptions<TInstance extends {}>
-  extends EntityOptions<TInstance> {
-  readonly pk:
-    | ((
-        value: TInstance,
-        parent?: any,
-        key?: string,
-      ) => string | number | undefined)
-    | keyof TInstance;
-}
-export default function EntitySchema<TBase extends Constructor>(
-  Base: TBase,
-  options?: EntityOptions<InstanceType<TBase>>,
-): any;
+
 export interface IEntityClass<TBase extends Constructor = any> {
   toJSON(): {
     name: string;
@@ -87,7 +36,9 @@ export interface IEntityClass<TBase extends Constructor = any> {
    * @param [args] ...args sent to Endpoint
    */
   pk<
-    T extends (new (...args: any[]) => IEntityInstance & InstanceType<TBase>) &
+    T extends (abstract new (
+      ...args: any[]
+    ) => IEntityInstance & InstanceType<TBase>) &
       IEntityClass &
       TBase,
   >(
@@ -113,20 +64,14 @@ export interface IEntityClass<TBase extends Constructor = any> {
     existing: any,
     incoming: any,
   ): boolean;
-  /** Determines the order of incoming entity vs entity already in store\
+  /** Determines the order of incoming entity vs entity already in store
    *
    * @see https://dataclient.io/docs/api/schema.Entity#shouldReorder
    * @returns true if incoming entity should be first argument of merge()
    */
   shouldReorder(
-    existingMeta: {
-      date: number;
-      fetchedAt: number;
-    },
-    incomingMeta: {
-      date: number;
-      fetchedAt: number;
-    },
+    existingMeta: { date: number; fetchedAt: number },
+    incomingMeta: { date: number; fetchedAt: number },
     existing: any,
     incoming: any,
   ): boolean;
@@ -161,11 +106,7 @@ export interface IEntityClass<TBase extends Constructor = any> {
       date: number;
       fetchedAt: number;
     },
-    incomingMeta: {
-      expiresAt: number;
-      date: number;
-      fetchedAt: number;
-    },
+    incomingMeta: { expiresAt: number; date: number; fetchedAt: number },
     existing: any,
     incoming: any,
   ): {
@@ -178,7 +119,9 @@ export interface IEntityClass<TBase extends Constructor = any> {
    * @param [props] Plain Object of properties to assign.
    */
   fromJS<
-    T extends (new (...args: any[]) => IEntityInstance & InstanceType<TBase>) &
+    T extends (abstract new (
+      ...args: any[]
+    ) => IEntityInstance & InstanceType<TBase>) &
       IEntityClass &
       TBase,
   >(
@@ -191,7 +134,9 @@ export interface IEntityClass<TBase extends Constructor = any> {
    * @see https://dataclient.io/rest/api/Entity#createIfValid
    */
   createIfValid<
-    T extends (new (...args: any[]) => IEntityInstance & InstanceType<TBase>) &
+    T extends (abstract new (
+      ...args: any[]
+    ) => IEntityInstance & InstanceType<TBase>) &
       IEntityClass &
       TBase,
   >(
@@ -210,8 +155,8 @@ export interface IEntityClass<TBase extends Constructor = any> {
     args: any[],
     visit: (...args: any) => any,
     addEntity: (...args: any) => any,
-    getEntity: GetEntity,
-    checkLoop: CheckLoop,
+    getEntity: (...args: any) => any,
+    checkLoop: (...args: any) => any,
   ): any;
   /** Do any transformations when first receiving input
    *
@@ -229,7 +174,9 @@ export interface IEntityClass<TBase extends Constructor = any> {
     getIndex: GetIndex,
   ): any;
   denormalize<
-    T extends (new (...args: any[]) => IEntityInstance & InstanceType<TBase>) &
+    T extends (abstract new (
+      ...args: any[]
+    ) => IEntityInstance & InstanceType<TBase>) &
       IEntityClass &
       TBase,
   >(
@@ -255,4 +202,54 @@ export interface IEntityInstance {
     args?: readonly any[],
   ): string | number | undefined;
 }
-export {};
+
+export type Constructor = abstract new (...args: any[]) => {};
+export type IDClass = abstract new (...args: any[]) => {
+  id: string | number | undefined;
+};
+export type PKClass = abstract new (...args: any[]) => {
+  pk(
+    parent?: any,
+    key?: string,
+    args?: readonly any[],
+  ): string | number | undefined;
+};
+// TODO: Figure out what Schema must be for each key
+type ValidSchemas<TInstance> = {
+  [k in keyof TInstance]?: Schema;
+};
+
+export type EntityOptions<TInstance extends {}> = {
+  readonly schema?: ValidSchemas<TInstance>;
+  readonly pk?:
+    | ((
+        value: TInstance,
+        parent?: any,
+        key?: string,
+      ) => string | number | undefined)
+    | keyof TInstance;
+  readonly key?: string;
+} & {
+  readonly [K in Extract<
+    keyof IEntityClass,
+    | 'process'
+    | 'merge'
+    | 'expiresAt'
+    | 'createIfValid'
+    | 'mergeWithStore'
+    | 'validate'
+    | 'shouldReorder'
+    | 'shouldUpdate'
+  >]?: IEntityClass<abstract new (...args: any[]) => TInstance>[K];
+};
+
+export interface RequiredPKOptions<TInstance extends {}>
+  extends EntityOptions<TInstance> {
+  readonly pk:
+    | ((
+        value: TInstance,
+        parent?: any,
+        key?: string,
+      ) => string | number | undefined)
+    | keyof TInstance;
+}
