+++ packages/endpoint/src/schemas/__tests__/EntitySchema.test.ts
@@ -349,7 +349,7 @@ describe(`${schema.Entity.name} normalization`, () => {
       MyEntity,
       { id: '1', title: 'second' },
       {},
-      { entities, entityMeta },
+      { entities, entityMeta, indexes: {} },
     ).entities;
     expect(entities.MyEntity['1']).toBeDefined();
     expect(entities.MyEntity['1']).toBe(secondEntities.MyEntity['1']);
