+++ website/src/components/Playground/editor-types/@data-client/normalizr.d.ts
@@ -160,9 +160,9 @@ interface Dep<Path, K = object> {
 }
 
 interface StoreData<E> {
-    entities?: Readonly<E>;
-    indexes?: Readonly<NormalizedIndex>;
-    entityMeta?: {
+    entities: Readonly<E>;
+    indexes: Readonly<NormalizedIndex>;
+    entityMeta: {
         readonly [entityKey: string]: {
             readonly [pk: string]: {
                 readonly date: number;
@@ -178,7 +178,7 @@ interface NormalizeMeta {
     fetchedAt?: number;
     args?: readonly any[];
 }
-declare const normalize: <S extends Schema = Schema, E extends Record<string, Record<string, any> | undefined> = Record<string, Record<string, any>>, R = NormalizeNullable<S>>(schema: S | undefined, input: any, { date, expiresAt, fetchedAt, args, }?: NormalizeMeta, { entities, indexes, entityMeta }?: StoreData<E>) => NormalizedSchema<E, R>;
+declare const normalize: <S extends Schema = Schema, E extends Record<string, Record<string, any> | undefined> = Record<string, Record<string, any>>, R = NormalizeNullable<S>>(schema: S | undefined, input: any, { date, expiresAt, fetchedAt, args, }?: NormalizeMeta, store?: StoreData<E>) => NormalizedSchema<E, R>;
 
 interface EntityCache {
     [key: string]: {
